chore: extract showing/hiding data selector logic to a hook
